+++ resolwe_bio/tests/processes/test_variant_calling.py
@@ -1,6 +1,6 @@
 # pylint: disable=missing-docstring
 import unittest
-from .utils import skipDockerFailure, BioProcessTestCase
+from resolwe_bio.utils.test import skipDockerFailure, BioProcessTestCase
 
 
 class VariantCallingTestCase(BioProcessTestCase):