+++ resolwe_bio/utils/__init__.py
@@ -1,0 +1,10 @@
+"""
+=========
+Utilities
+=========
+
+.. automodule:: resolwe_bio.utils.test
+   :members:
+
+"""
+from __future__ import absolute_import, division, print_function, unicode_literals